Fixes GetAllSessions() reliability, refs #1
Using ctypes.cast() wasn't incrementing the AddRef() reference counter.
However the reference was decremented twice via Release(),
once for the ctl (IUknown) object release and once for the
mgr (IAudioSessionManager2) object release.
But since casting to the IAudioSessionManager2 wasn't incrementing
the ref counter, it lead to a memory corruption.
Same thing happened later in the code with the ctl2 object.