:construction_worker: Fix CI virtual audio install
Change the system date as a workaround for the expired Scream certs.
The job was driver install job was hanging, the error log was:
```
Updating drivers for *Scream from D:\a\pycaw\pycaw\Scream\Install\driver\x64\Scream.inf.
Error: The operation was canceled.
```
Refs: https://github.com/duncanthrax/scream/issues/202
Also bump Windows and Scream versions.